Add files via upload
Added the files used for this group project. The PowerPoint presentation shows the analysis done using Tableau, a group report constructed using Microsoft Word, and the chosen dataset, which is called "America Population Data 1955-2020". This dataset can be located on Kaggle. The "Population Forecast" file contains similar data to the previous dataset but includes the years 2020-2040. The "Cities by Population" dataset contains a list of the cities and their populations. 


Link to dataset: https://www.kaggle.com/datasets/haseebchohan/america-population-data-1955-2020